Fix LE device not working on Ventura
Look like starting from Ventura, the bluetooth daemon expects an
LE PHY Update Complete event to be triggered from the controller
before starting the connection. However for Intel cards, that event
is not sent by the firmware, so we can fake it by issuing an extra
LE command and modify its response.